Bugs fixed, HTTP error handler page added

+++ app/Http/Controllers/Backend/AttributeGroupController.php
@@ -21,6 +21,9 @@ class AttributeGroupController extends Controller
      */
     public function index()
     {
+        if (!$this->isDatabaseConnected()) {
+            abort(500, 'Database Connection Error');
+        }
         $attributesGroup = AttributeGroup::all();
         return view('admin.attributes.index', compact('attributesGroup'));
     }
@@ -44,6 +47,9 @@ class AttributeGroupController extends Controller
      */
     public function store(CreateAttributeGroupRequest $request)
     {
+        if (!$this->isDatabaseConnected()) {
+            abort(500, 'Database Connection Error');
+        }
         $attributeGroup = new AttributeGroup();
         $attributeGroup->title = $request->input('title');
         $attributeGroup->type = $request->input('type');
@@ -71,6 +77,9 @@ class AttributeGroupController extends Controller
      */
     public function edit($id)
     {
+        if (!$this->isDatabaseConnected()) {
+            abort(500, 'Database Connection Error');
+        }
         $attributesGroup = AttributeGroup::findOrFail($id);
         return view('admin.attributes.edit', compact('attributesGroup'));
     }
@@ -84,6 +93,9 @@ class AttributeGroupController extends Controller
      */
     public function update(CreateAttributeGroupRequest $request, $id)
     {
+        if (!$this->isDatabaseConnected()) {
+            abort(500, 'Database Connection Error');
+        }
         $attributeGroup = AttributeGroup::findOrFail($id);
         $attributeGroup->title = $request->title;
         $attributeGroup->type = $request->type;
@@ -100,9 +112,15 @@ class AttributeGroupController extends Controller
      */
     public function destroy($id)
     {
+        if (!$this->isDatabaseConnected()) {
+            abort(500, 'Database Connection Error');
+        }
         $attributeGroup = AttributeGroup::findOrFail($id);
+        foreach ($attributeGroup->attributesValue as $item) {
+            $item->delete();
+        }
         $attributeGroup->delete();
-        Session::flash('attributes', 'ویژگی با موفقیت حذف شد!');
+        Session::flash('attributes', 'ویژگی و مقادیر آن با موفقیت حذف شدند!');
         return redirect('/administrator/attributes-group');
     }
 }
